軽微

+++ Engine/util/Util.cpp
@@ -4,7 +4,7 @@
 #include <time.h>
 
 static std::mt19937& getRand() {
-	static std::mt19937 mt(time(0));
+	static std::mt19937 mt(static_cast<UINT>(time(0)));
 	return mt;
 }
 
@@ -13,7 +13,7 @@ int Util::RandomInt(int begine, int end) {
 	return l_rand(getRand());
 }
 
-int Util::RandomFloat(float begine, float end) {
+float Util::RandomFloat(float begine, float end) {
 	std::uniform_real_distribution<float> l_rand(begine, end);
 	return l_rand(getRand());
 }
